requirements.txt
Django==2.1.7
django-appconf==1.0.2
django-bootstrap4==0.0.7
django-compress==1.0.1
django-countries==5.3.3
django-recaptcha==2.0.2
django-sass-processor==0.7.2
djangorestframework==3.9.1
gunicorn==19.9.0
libsass==0.17.0
Pillow==5.4.1
psycopg2-binary==2.7.7
pytz==2018.9
rcssmin==1.0.6
rjsmin==1.0.12
six==1.12.0
